backend/tile-server/pyproject.toml
[tool.poetry]
name = "tile-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "tile_server"}]

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^9.3.0"
numpy = "^1.25.1"
flask = "^2.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
